# view/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "MVC",
    "title": "MilkingCow Farm"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/searchCow": {
      "post": {
        "tags": [
          "searchCow"
        ],
        "summary": "",
        "description": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "idCow",
            "description": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "Success"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "Something Wrong"
              }
            }
          }
        }
      }
    },
    "/MilkingWhiteCow": {
      "post": {
        "tags": [
          "MilkingWhiteCow"
        ],
        "summary": "",
        "description": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "idCow",
            "description": "",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "lemonEated",
            "description": "",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "MilkingWhiteCow Successful"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "MilkingWhiteCow UnSuccessful"
              }
            }
          }
        }
      }
    },
    "/MilkingBrownCow": {
      "post": {
        "tags": [
          "MilkingBrownCow"
        ],
        "summary": "",
        "description": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "idCow",
            "description": "",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "MilkingBrownCow Successful"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "MilkingBrownCow UnSuccessful"
              }
            }
          }
        }
      }
    },
    "/resetBSOD": {
      "get": {
        "tags": [
          "resetBSOD"
        ],
        "summary": "",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "Success"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "example": {
                "message": "Something Wrong"
              }
            }
          }
        }
      }
    }
  }
}